refactor: implement FastAPI dependency injection for data layer
- Create RecipeRepository abstract base class and InMemoryRecipeRepository implementation
- Add dependency injection functions in app/dependencies.py
- Update RecipeService to accept repository as dependency
- Modify API endpoints to use Depends() for dependency injection
- Update tests to work with new DI architecture using dependency overrides
- Maintain all existing API functionality while improving architecture
- Prepare codebase for easy database swapping in the future
