+++ app/repositories/__init__.py
@@ -1,0 +1,1 @@
+# Repository module for data layer abstraction
