+++ main.py
@@ -1,4 +1,7 @@
-from app.core.app import app
+from app.core.app import create_app
+
+# Create the FastAPI application instance
+app = create_app()
 
 # This file serves as the entry point for the application
 # The actual FastAPI app is now defined in app/core/app.py
