+++ test_ping.py
@@ -1,7 +1,8 @@
 # test_ping.py
 from fastapi.testclient import TestClient
-from main import app
+from app.core.app import create_app
 
+app = create_app()
 client = TestClient(app)
 
 def test_ping():
